refactor: move modules to `common`
